Working issues 6

+++ app/src/main/java/com/example/aas_app/data/dao/EvaluationResultDao.kt
@@ -16,7 +16,7 @@ interface EvaluationResultDao {
     fun getEvaluationResultsForStudent(studentId: Long): Flow<List<PeclEvaluationResultEntity>>
 
     @Query("SELECT AVG(score) FROM pecl_evaluation_results WHERE student_id = :studentId AND question_id IN (SELECT q.id FROM pecl_questions q JOIN question_assignments a ON q.id = a.question_id JOIN pecl_tasks t ON a.task_id = t.id JOIN pecl_pois p ON t.poi_id = p.id WHERE p.id = :poiId)")
-    suspend fun getAverageScoreForStudent(studentId: Long, poiId: Long): Float
+    suspend fun getAverageScoreForStudent(studentId: Long, poiId: Long): Double
 
     @Query("SELECT * FROM pecl_evaluation_results WHERE question_id = :questionId")
     suspend fun getEvaluationsForQuestion(questionId: Long): List<PeclEvaluationResultEntity>
